Exdir

Documentation for Exdir.

Exdir.Raw

Exdir.create_group

Exdir.create_group

Exdir.create_object_directory

Exdir.exdiropen

Exdir.form_location

Exdir.is_exdir_object

Exdir.require_group

Exdir.Raw — Type

Raw objects are simple folders with any content.

Raw objects currently have no features apart from showing their path.

Exdir.create_group — Method

create_group(file, name)

Exdir.create_group — Method

create_group(grp, name)

Exdir.create_object_directory — Method

create_object_directory(directory, metadata)

Create object directory and meta file if the directory doesn't already exist.

TODO say something about expected correctness of metadata dict.

Exdir.exdiropen — Method

exdiropen(directory::AbstractString, mode::AbstractString)

Opens an ExDir tree at path directory.

"r": Open for reading only, failing if no file exists "r+": Open for reading and writing, failing if no file exists "w"/"w+": Open for reading and writing, overwriting the file if it already exists "a"/"a+": Open for reading and writing, creating a new file if none exists, but preserving the existing file if one is present

Exdir.form_location — Method

form_location(directory)

Prepare the given directory location for use by removing any trailing slashes/directory signifiers, ensuring the directory ends with the .exdir extension, and making the directory path absolute.

Exdir.is_exdir_object — Method

WARNING: Does not test of inside exdir directory, only if the object can be an exdir object (i.e. a directory).

Exdir.require_group — Method

require_group(grp, name)